requirements.txt
# Merged from AdOps, QA-Agent, AI-Stuff (Wake Lock QA Bot). No streamlit.
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
numpy>=1.24.0
slack-bolt>=1.18.0
python-dotenv>=1.0.0
requests>=2.31.0
androguard>=4.1.2
google-play-scraper>=1.2.4
beautifulsoup4>=4.12.0
lxml>=5.0.0
playwright>=1.40.0
flask>=3.0.0
PyQt5>=5.15.0
